refactor(#5015): split TUI retry helpers below giant cap
